Implement Keycloak (OIDC) integration #16816
Wrote a Gin middleware to handle authentication. It extracts the token and verifies it by calling Keycloak. It caches verified tokens.

The middleware extracts the payload of the token and passes in Gin context for the web services to use it if required.

With this the web services are not required to handle the authentication part.